server/requirements.txt
aiohttp==3.7.4.post0
aiohttp-index==0.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
black==20.8b1
chardet==4.0.0
click==7.1.2
idna==3.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.3
pathspec==0.8.1
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
regex==2021.3.17
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
yarl==1.6.3
